pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-accel-api-platform"
version = "0.1.0"
description = "AI-accelerated API platform with GPU/MPS/CPU auto-selection and vector search."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "sqlalchemy[asyncio]",
  "asyncpg",
  "psycopg[binary]",
  "alembic",
  "pgvector",
  "redis",
  "rq",
  "sentence-transformers",
  "torch",
  "numpy",
  "orjson",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "bcrypt<4",
  "python-multipart",
  "prometheus-client",
  "structlog",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "httpx",
  "pre-commit",
  "types-redis",
  "types-python-jose",
]
grpc = [
  "grpcio",
  "grpcio-tools",
]
otel = [
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-logging",
]
cpp = [
  "scikit-build-core",
  "pybind11",
  "cmake",
  "ninja",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = []
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = [
  "torch.*",
  "sentence_transformers.*",
  "redis.*",
  "rq.*",
  "pgvector.*",
  "opentelemetry.*",
  "ai_accel_api_platform.cpp.*",
  "orjson",
  "passlib.*",
  "prometheus_client.*",
  "jose.*",
  "grpc.*",
  "grpc_tools.*",
  "structlog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
asyncio_mode = "auto"
markers = ["integration: marks tests that hit external services"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
